backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
PyMuPDF
python-multipart